src/requirements.txt
torch>=2.0.0
monai[all]>=1.3.0
nibabel>=5.0.0
scipy>=1.10.0
scikit-image>=0.21.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
